Pull up HamburgerMenu too Report.js.
