+++ components/frontend/src/subject/Subjects.test.js
@@ -40,6 +40,7 @@ function renderSubjects(permissions = [], initialEntries = []) {
                     report={report}
                     reports={[report]}
                     tags={[]}
+                    visibleDetailsTabs={[]}
                 />
             </DataModel.Provider>
         </Permissions.Provider>
@@ -83,130 +84,3 @@ it("moves a subject", async () => {
     await act(async () => fireEvent.click(screen.getByText("dummy option 2")))
     expect(fetch_server_api.fetch_server_api).toHaveBeenLastCalledWith("post", "subject/undefined/move/undefined", { });
 });
-
-it("hides metrics not requiring action", async () => {
-    renderSubjects();
-    await act(async () => fireEvent.click(screen.getAllByTestId("HamburgerMenu")[0]));
-    expect(screen.getAllByText(/Hide metrics not requiring action/).length).toBe(1)
-    await act(async () => fireEvent.click(screen.getAllByText(/Hide metrics not requiring action/)[0]));
-    expect(screen.queryAllByText(/Hide metrics not requiring action/).length).toBe(0)
-    await act(async () => fireEvent.click(screen.getAllByText(/Show all metrics/)[0]))
-    expect(screen.getAllByText(/Hide metrics not requiring action/).length).toBe(1)
-})
-
-it('sorts the metrics by name', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-})
-
-it('sorts the metrics by status', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Status/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Status/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by measurement value', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Measurement/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Measurement/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by target value', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Target/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Target/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by source', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Source/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Source/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by comment', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Comment/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Comment/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by issues', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Issues/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Issues/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by tags', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Tags/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Tags/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('sorts the metrics by unit', () => {
-    fetch_server_api.fetch_server_api = jest.fn().mockReturnValue({ then: jest.fn().mockReturnValue({ finally: jest.fn() }) });
-    renderSubjects([], ["?nr_dates=2"]);
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Unit/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Unit/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('keeps sort direction when sorting another column', () => {
-    renderSubjects();
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-    fireEvent.click(screen.getAllByText(/Status/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
-
-it('stops sorting when add metric is clicked', async () => {
-    renderSubjects([EDIT_REPORT_PERMISSION]);
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-    await act(async () => fireEvent.click(screen.getAllByText(/Add metric/)[0]))
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-})
-
-it('stops sorting when metrics are reordered', async () => {
-    fetch_server_api.fetch_server_api = jest.fn().mockReturnValue({ then: jest.fn().mockReturnValue({ finally: jest.fn() }) });
-    renderSubjects([EDIT_REPORT_PERMISSION]);
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M1", "M2"])
-    fireEvent.click(screen.getAllByText(/Metric/)[0])
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-    await act(async () => fireEvent.click(screen.getAllByTitle(/expand/)[1]))
-    await act(async () => fireEvent.click(screen.getAllByRole("button", {text: "Move metric to the last row"})[0]))
-    expect(screen.queryAllByText(/M[12]/).map((element) => element.innerHTML)).toMatchObject(["M2", "M1"])
-})
